+++ app/src/main/java/com/kionavani/todotask/domain/TodoItemsRepository.kt
@@ -1,7 +1,5 @@
 package com.kionavani.todotask.domain
 
-import com.kionavani.todotask.data.ToDoItem
-import com.kionavani.todotask.data.remote.NetworkResult
 import kotlinx.coroutines.flow.StateFlow
 
 /**
